+++ lib/generator.js
@@ -4,7 +4,7 @@ const xfs = require('fs.extra');
 const minimatch = require('minimatch');
 const parser = require('@asyncapi/parser');
 const {configureReact, renderReact} = require('./renderer/react');
-const {configureNunjucks, renderNunjucks} = require('./renderer/nunjucks');
+const {configureNunjucks, renderNunjucks, saveRenderedReactContent} = require('./renderer/nunjucks');
 const { parse, AsyncAPIDocument } = parser;
 const ramlDtParser = require('@asyncapi/raml-dt-schema-parser');
 const openapiSchemaParser = require('@asyncapi/openapi-schema-parser');
@@ -588,29 +588,11 @@ class Generator {
    */
   async renderAndWriteToFile(asyncapiDocument, templateFilePath, outputpath, extraTemplateData) {
     const renderContent =  await this.renderFile(asyncapiDocument, templateFilePath, extraTemplateData);
-    if(renderContent !== undefined){
-      if (isReactTemplate(this.templateConfig)) {
-        let filePath = outputpath;
-        let permissions = 0o666;
-        const content = renderContent.content;
-        if (content === undefined || typeof content !== 'string') throw Error(`Template file not rendered correctly, was rendered as ${content}`);
-        if (renderContent.metadata !== undefined) {
-          if (renderContent.metadata.permissions !== undefined) {
-            permissions = renderContent.metadata.permissions;
-          }
-          if (renderContent.metadata.fileName !== undefined) {
-            const newFileName = renderContent.metadata.fileName.trim();
-            const basepath = path.basename(filePath);
-            filePath = filePath.replace(basepath, newFileName);
-          }
-        }
-
-        await writeFile(filePath, content, {
-          mode: permissions
-        });
-      } else {
-        await writeFile(outputpath, renderContent);
-      }
+    if (renderContent !== undefined) return; 
+    if (isReactTemplate(this.templateConfig)) {
+      await saveRenderedReactContent(renderContent, outputpath);
+    } else {
+      await writeFile(outputpath, renderContent);
     }
   }
