+++ lib/renderer/react.js
@@ -1,5 +1,8 @@
 const path = require('path');
 const AsyncReactSDK = require('asyncAPISDK');
+const {
+  writeFile
+} = require('../utils');
 
 const reactExport = module.exports;
 
@@ -44,4 +47,30 @@ reactExport.renderReact = async (asyncapiDocument, filePath, extraTemplateData =
       ...extraTemplateData
     }, 
     debug);
+};
+
+/**
+ * Save the rendered react content based on the meta data available.
+ * 
+ * @param {TemplateRenderResult} renderContent the react content rendered
+ * @param {String} outputpath Path to the file being rendered.
+ */
+reactExport.saveRenderedReactContent = async (renderContent, outputpath) => {
+  let filePath = outputpath;
+  let permissions = 0o666;
+  const content = renderContent.content;
+  if (content === undefined || typeof content !== 'string') throw Error(`Template file not rendered correctly, was rendered as ${content}`);
+  if (renderContent.metadata !== undefined) {
+    if (renderContent.metadata.permissions !== undefined) {
+      permissions = renderContent.metadata.permissions;
+    }
+    if (renderContent.metadata.fileName !== undefined) {
+      const newFileName = renderContent.metadata.fileName.trim();
+      const basepath = path.basename(filePath);
+      filePath = filePath.replace(basepath, newFileName);
+    }
+  }
+  await writeFile(filePath, content, {
+    mode: permissions
+  });
 };
